# tests/fixtures/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "CombinedRequest": {
        "properties": {
          "target_tone": {
            "default": "professional",
            "title": "Target Tone",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "user_instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Instruction"
          }
        },
        "required": [
          "text"
        ],
        "title": "CombinedRequest",
        "type": "object"
      },
      "CombinedResponse": {
        "properties": {
          "prediction": {
            "$ref": "#/components/schemas/PredictResponse"
          },
          "rewrite": {
            "$ref": "#/components/schemas/RewriteResponse"
          },
          "rewritten_prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "prediction",
          "rewrite"
        ],
        "title": "CombinedResponse",
        "type": "object"
      },
      "EmotionScore": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "label",
          "score"
        ],
        "title": "EmotionScore",
        "type": "object"
      },
      "ExplainToken": {
        "properties": {
          "importance": {
            "title": "Importance",
            "type": "number"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "importance"
        ],
        "title": "ExplainToken",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MonitoringSnapshotResponse": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "distribution": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Distribution",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "window_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Window Counts",
            "type": "object"
          },
          "window_distribution": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Window Distribution",
            "type": "object"
          },
          "window_size": {
            "title": "Window Size",
            "type": "integer"
          },
          "window_total": {
            "title": "Window Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "counts",
          "distribution",
          "window_total",
          "window_counts",
          "window_distribution",
          "window_size"
        ],
        "title": "MonitoringSnapshotResponse",
        "type": "object"
      },
      "PredictRequest": {
        "properties": {
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "top_k": {
            "default": 3,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "text"
        ],
        "title": "PredictRequest",
        "type": "object"
      },
      "PredictResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "emotion": {
            "title": "Emotion",
            "type": "string"
          },
          "explanation_tokens": {
            "items": {
              "$ref": "#/components/schemas/ExplainToken"
            },
            "title": "Explanation Tokens",
            "type": "array"
          },
          "low_confidence": {
            "default": false,
            "title": "Low Confidence",
            "type": "boolean"
          },
          "model_source": {
            "default": "lightweight",
            "title": "Model Source",
            "type": "string"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/EmotionScore"
            },
            "title": "Scores",
            "type": "array"
          },
          "sentence_distributions": {
            "items": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "title": "Sentence Distributions",
            "type": "array"
          },
          "sentiment_percentages": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Sentiment Percentages",
            "type": "object"
          }
        },
        "required": [
          "emotion",
          "confidence",
          "scores",
          "sentiment_percentages",
          "explanation_tokens"
        ],
        "title": "PredictResponse",
        "type": "object"
      },
      "RewriteRequest": {
        "properties": {
          "target_tone": {
            "default": "professional",
            "title": "Target Tone",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "user_instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Instruction"
          }
        },
        "required": [
          "text"
        ],
        "title": "RewriteRequest",
        "type": "object"
      },
      "RewriteResponse": {
        "properties": {
          "applied_tone": {
            "title": "Applied Tone",
            "type": "string"
          },
          "detected_emotion": {
            "title": "Detected Emotion",
            "type": "string"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "original_text": {
            "title": "Original Text",
            "type": "string"
          },
          "quality_flags": {
            "items": {
              "type": "string"
            },
            "title": "Quality Flags",
            "type": "array"
          },
          "rewrite_confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Rewrite Confidence",
            "type": "number"
          },
          "rewrite_source": {
            "default": "lightweight",
            "title": "Rewrite Source",
            "type": "string"
          },
          "rewritten_text": {
            "title": "Rewritten Text",
            "type": "string"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "original_text",
          "detected_emotion",
          "rewritten_text",
          "applied_tone",
          "suggestions",
          "rewrite_confidence"
        ],
        "title": "RewriteResponse",
        "type": "object"
      },
      "TrainRequest": {
        "properties": {
          "base_model": {
            "default": "distilbert-base-uncased",
            "title": "Base Model",
            "type": "string"
          },
          "batch_size": {
            "default": 8,
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Batch Size",
            "type": "integer"
          },
          "epochs": {
            "default": 1,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Epochs",
            "type": "integer"
          },
          "learning_rate": {
            "default": 2e-05,
            "exclusiveMinimum": 0.0,
            "title": "Learning Rate",
            "type": "number"
          },
          "max_length": {
            "default": 128,
            "maximum": 512.0,
            "minimum": 32.0,
            "title": "Max Length",
            "type": "integer"
          },
          "train_csv_path": {
            "title": "Train Csv Path",
            "type": "string"
          },
          "val_csv_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Val Csv Path"
          }
        },
        "required": [
          "train_csv_path"
        ],
        "title": "TrainRequest",
        "type": "object"
      },
      "TrainResponse": {
        "properties": {
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Metrics",
            "type": "object"
          },
          "num_labels": {
            "title": "Num Labels",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "num_labels",
          "labels",
          "metrics"
        ],
        "title": "TrainResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Transformer Sentiment Transformation API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "root__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Root  Get",
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/analyze-rewrite": {
      "post": {
        "operationId": "analyze_and_rewrite_analyze_rewrite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CombinedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombinedResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "readiness_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Readiness Health Ready Get",
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/monitoring/snapshot": {
      "get": {
        "operationId": "monitoring_snapshot_monitoring_snapshot_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringSnapshotResponse"
                }
              }
            }
          }
        }
      }
    },
    "/predict": {
      "post": {
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rewrite": {
      "post": {
        "operationId": "rewrite_rewrite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RewriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewriteResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tones/suggestions": {
      "get": {
        "operationId": "tone_suggestions_tones_suggestions_get",
        "parameters": [
          {
            "in": "query",
            "name": "emotion",
            "required": true,
            "schema": {
              "title": "Emotion",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Tone Suggestions Tones Suggestions Get",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/train": {
      "post": {
        "operationId": "train_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  }
}